+++ tests/testthat/test_methods-Chromatograms.R
@@ -247,14 +247,35 @@ test_that("isEmpty,Chromatograms works", {
     chrs <- Chromatograms(list(ch, ch1, ch2, ch3), nrow = 2)
 
     expect_true(!isEmpty(chrs))
-
+    plot(chrs)
+    
     chrs <- Chromatograms()
     expect_true(isEmpty(chrs))
+    expect_warning(plot(chrs))
     
     ints <- rep(NA_real_, 105)
     ch1 <- Chromatogram(rtime = 1:length(ints), ints)
     ints <- rep(NA_real_, 64)
     ch2 <- Chromatogram(rtime = 1:length(ints), ints)
     chrs <- Chromatograms(list(ch1, ch2), nrow = 2)
-    expect_true(isEmpty(chrs))    
+    expect_true(isEmpty(chrs))
+    expect_warning(plot(chrs))
+    
+    ## Only one row is empty.
+    ints <- rep(NA_real_, 105)
+    ch1 <- Chromatogram(rtime = 1:length(ints), ints)
+    ints <- abs(rnorm(64))
+    ch2 <- Chromatogram(rtime = 1:length(ints), ints)
+    chrs <- Chromatograms(list(ch1, ch2), nrow = 2)
+    expect_true(!isEmpty(chrs))
+    expect_warning(plot(chrs))
+
+    ## 2x2 first row NA
+    chrs <- Chromatograms(list(ch1, ch2, ch1, ch2), nrow = 2)
+    expect_warning(plot(chrs))
+
+    ## 2x2 first col NA
+    chrs <- Chromatograms(list(ch1, ch1, ch2, ch2), nrow = 2)
+    expect_true(!isEmpty(chrs))
+    plot(chrs)
 })