+++ tests/testthat/test_methods-Chromatogram.R
@@ -47,6 +47,17 @@ test_that("Chromatogram accessors", {
     expect_equal(productMz(chr), 432)
     chr <- Chromatogram(intensity = int, rt = rt, productMz = 432)
     expect_equal(productMz(chr), c(432, 432))
+
+    ## msLevel
+    chr@msLevel <- 2L
+    expect_equal(msLevel(chr), 2L)
+    expect_true(validObject(chr))
+    chr@msLevel <- 1:4
+    expect_equal(msLevel(chr), 1:4)
+    expect_true(validObject(chr))
+
+    chr <- Chromatogram(intensity = int, rt = rt, msLevel = 4)
+    expect_equal(msLevel(chr), 4L)
 })
 
 test_that("clean,Chromatogram works", {
@@ -109,4 +120,21 @@ test_that("filterRt,Chromatogram works", {
     expect_true(length(chr_2) == 0)
     expect_equal(intensity(chr_2), numeric())
     expect_equal(rtime(chr_2), numeric())
+})
+
+test_that("isEmpty,Chromatogram and plot,Chromatogram work", {
+    chr <- Chromatogram()
+    expect_true(isEmpty(chr))
+    expect_warning(plot(chr))
+    
+    int <- rnorm(100, mean = 200, sd = 2)
+    rt <- rnorm(100, mean = 300, sd = 3)
+    chr <- Chromatogram(intensity = int, rtime = sort(rt))
+    expect_true(!isEmpty(chr))
+    plot(chr)
+    
+    chr <- Chromatogram(intensity = rep_len(NA_real_, length(rt)),
+                        rtime = sort(rt))
+    expect_true(isEmpty(chr))
+    expect_warning(plot(chr))
 })